Moves porosity functions in material properties
