+++ include/private/tdymaterialpropertiesimpl.h
@@ -7,6 +7,7 @@ PETSC_INTERN void TDySoilDensityFunctionDefault(PetscReal*,PetscReal*);
 PETSC_INTERN void TDySpecificSoilHeatCapacityFunctionDefault(PetscReal*,PetscReal*);
 PETSC_INTERN PetscErrorCode TDyPermeabilityFunctionDefault(TDy,double*,double*,void*);
 PETSC_INTERN PetscErrorCode TDyThermalConductivityFunctionDefault(TDy,double*,double*,void*);
+PETSC_INTERN PetscErrorCode TDyPorosityFunctionDefault(TDy,double*,double*,void*);
 
 #endif
