--- include/private/tdyporosityimpl.h
@@ -1,9 +1,0 @@
-#if !defined(TDYPOROSITYIMPL_H)
-#define TDYPOROSITYIMPL_H
-
-#include <petsc.h>
-#include <tdycore.h>
-
-PETSC_INTERN PetscErrorCode TDyPorosityFunctionDefault(TDy,double*,double*,void*);
-
-#endif
