+++ src/tdymaterialproperties.c
@@ -12,6 +12,13 @@ PetscErrorCode TDySetPermeabilityFunction(TDy tdy, PetscErrorCode(*f)(TDy,PetscR
   PetscFunctionReturn(0);
 }
 
+PetscErrorCode TDySetPorosityFunction(TDy tdy, PetscErrorCode(*f)(TDy,PetscReal*,PetscReal*,void*),void *ctx) {
+  PetscFunctionBegin;
+  if (f) tdy->ops->computeporosity = f;
+  if (ctx) tdy->porosityctx = ctx;
+  PetscFunctionReturn(0);
+}
+
 PetscErrorCode TDySetThermalConductivityFunction(TDy tdy, PetscErrorCode(*f)(TDy,PetscReal*,PetscReal*,void*),void *ctx) {
   PetscFunctionBegin;
   if (f) tdy->ops->computethermalconductivity = f;
@@ -88,6 +95,21 @@ PetscErrorCode TDySetPermeabilityTensor(TDy tdy,SpatialFunction f) {
   PetscFunctionReturn(0);
 }
 
+PetscErrorCode TDySetPorosity(TDy tdy,SpatialFunction f) {
+  PetscInt dim,c,cStart,cEnd;
+  PetscErrorCode ierr;
+  PetscFunctionBegin;
+  TDY_START_FUNCTION_TIMER()
+  PetscValidPointer(tdy,1);
+  PetscValidPointer(f,2);
+  ierr = DMGetDimension(tdy->dm,&dim); CHKERRQ(ierr);
+  ierr = DMPlexGetHeightStratum(tdy->dm,0,&cStart,&cEnd); CHKERRQ(ierr);
+  ierr = PetscMemzero(tdy->matprop_porosity,sizeof(PetscReal)*(cEnd-cStart)); CHKERRQ(ierr);
+  for(c=cStart; c<cEnd; c++) f(&(tdy->X[dim*c]),&(tdy->matprop_porosity[c-cStart]));
+  TDY_STOP_FUNCTION_TIMER()
+  PetscFunctionReturn(0);
+}
+
 PetscErrorCode TDySetSoilSpecificHeatCapacity(TDy tdy,SpatialFunction f) {
   PetscInt dim,c,cStart,cEnd;
   PetscErrorCode ierr;
@@ -153,6 +175,26 @@ PetscErrorCode TDySetCellPermeability(TDy tdy,PetscInt c,PetscReal *K) {
   PetscFunctionReturn(0);
 }
 
+PetscErrorCode TDySetPorosityValuesLocal(TDy tdy, PetscInt ni, const PetscInt ix[], const PetscScalar y[]){
+
+  PetscInt i;
+
+  PetscFunctionBegin;
+  TDY_START_FUNCTION_TIMER()
+  if (!ni) PetscFunctionReturn(0);
+
+  for(i=0; i<ni; i++) {
+    tdy->matprop_porosity[ix[i]] = y[i];
+  }
+
+  TDY_STOP_FUNCTION_TIMER()
+  PetscFunctionReturn(0);
+}
+
+/*
+  Get material properties cell-by-cell
+*/
+
 PetscErrorCode TDyGetBlockPermeabilityValuesLocal(TDy tdy, PetscInt *ni, PetscScalar y[]){
 
   PetscInt c,cStart,cEnd,j;
@@ -178,6 +220,30 @@ PetscErrorCode TDyGetBlockPermeabilityValuesLocal(TDy tdy, PetscInt *ni, PetscSc
     }
   }
 
+  PetscFunctionReturn(0);
+}
+
+PetscErrorCode TDyGetPorosityValuesLocal(TDy tdy, PetscInt *ni, PetscScalar y[]){
+
+  PetscInt c,cStart,cEnd;
+  PetscInt junkInt, gref;
+  PetscErrorCode ierr;
+
+  PetscFunctionBegin;
+  TDY_START_FUNCTION_TIMER()
+
+  ierr = DMPlexGetHeightStratum(tdy->dm,0,&cStart,&cEnd); CHKERRQ(ierr);
+  *ni = 0;
+
+  for (c=cStart; c<cEnd; c++) {
+    ierr = DMPlexGetPointGlobal(tdy->dm,c,&gref,&junkInt); CHKERRQ(ierr);
+    if (gref>=0) {
+      y[*ni] = tdy->matprop_porosity[c-cStart];
+      *ni += 1;
+    }
+  }
+
+  TDY_STOP_FUNCTION_TIMER()
   PetscFunctionReturn(0);
 }
 
@@ -222,5 +288,11 @@ PetscErrorCode TDyThermalConductivityFunctionDefault(TDy tdy, double *x, double 
   for (int i=0; i<dim; i++)
     K[i*dim+i] = 1.;
   return 0;
+  PetscFunctionReturn(0);
+}
+
+PetscErrorCode TDyPorosityFunctionDefault(TDy tdy, double *x, double *por, void *ctx) {
+  PetscFunctionBegin;
+  *por = 0.25;
   PetscFunctionReturn(0);
 }
