--- src/tdyporosity.c
@@ -1,70 +1,0 @@
-#include <private/tdycoreimpl.h>
-#include <tdytimers.h>
-
-PetscErrorCode TDySetPorosity(TDy tdy,SpatialFunction f) {
-  PetscInt dim,c,cStart,cEnd;
-  PetscErrorCode ierr;
-  PetscFunctionBegin;
-  TDY_START_FUNCTION_TIMER()
-  PetscValidPointer(tdy,1);
-  PetscValidPointer(f,2);
-  ierr = DMGetDimension(tdy->dm,&dim); CHKERRQ(ierr);
-  ierr = DMPlexGetHeightStratum(tdy->dm,0,&cStart,&cEnd); CHKERRQ(ierr);
-  ierr = PetscMemzero(tdy->matprop_porosity,sizeof(PetscReal)*(cEnd-cStart)); CHKERRQ(ierr);
-  for(c=cStart; c<cEnd; c++) f(&(tdy->X[dim*c]),&(tdy->matprop_porosity[c-cStart]));
-  TDY_STOP_FUNCTION_TIMER()
-  PetscFunctionReturn(0);
-}
-
-PetscErrorCode TDySetPorosityFunction(TDy tdy, PetscErrorCode(*f)(TDy,PetscReal*,PetscReal*,void*),void *ctx) {
-  PetscFunctionBegin;
-  if (f) tdy->ops->computeporosity = f;
-  if (ctx) tdy->porosityctx = ctx;
-  PetscFunctionReturn(0);
-}
-
-PetscErrorCode TDySetPorosityValuesLocal(TDy tdy, PetscInt ni, const PetscInt ix[], const PetscScalar y[]){
-
-  PetscInt i;
-
-  PetscFunctionBegin;
-  TDY_START_FUNCTION_TIMER()
-  if (!ni) PetscFunctionReturn(0);
-
-  for(i=0; i<ni; i++) {
-    tdy->matprop_porosity[ix[i]] = y[i];
-  }
-
-  TDY_STOP_FUNCTION_TIMER()
-  PetscFunctionReturn(0);
-}
-
-PetscErrorCode TDyGetPorosityValuesLocal(TDy tdy, PetscInt *ni, PetscScalar y[]){
-
-  PetscInt c,cStart,cEnd;
-  PetscInt junkInt, gref;
-  PetscErrorCode ierr;
-
-  PetscFunctionBegin;
-  TDY_START_FUNCTION_TIMER()
-
-  ierr = DMPlexGetHeightStratum(tdy->dm,0,&cStart,&cEnd); CHKERRQ(ierr);
-  *ni = 0;
-
-  for (c=cStart; c<cEnd; c++) {
-    ierr = DMPlexGetPointGlobal(tdy->dm,c,&gref,&junkInt); CHKERRQ(ierr);
-    if (gref>=0) {
-      y[*ni] = tdy->matprop_porosity[c-cStart];
-      *ni += 1;
-    }
-  }
-
-  TDY_STOP_FUNCTION_TIMER()
-  PetscFunctionReturn(0);
-}
-
-PetscErrorCode TDyPorosityFunctionDefault(TDy tdy, double *x, double *por, void *ctx) {
-  PetscFunctionBegin;
-  *por = 0.25;
-  PetscFunctionReturn(0);
-}
